python-publisher/requirements.txt
essential_generators
kafka-python
requests
pandas
